fix(benchmarks): attempt to fix the generation of benchmarks output markdown file
