+++ benches/run_benches.py
@@ -2,42 +2,7 @@ import json
 import subprocess
 from pathlib import Path
 
-import pandas as pd
-
-
-def compare_benchmarks():
-    benches_dir = Path(__file__).parent
-    results_path = benches_dir / "results.json"
-    with open(results_path) as f:
-        data = json.load(f)
-
-    df = pd.json_normalize(data)
-
-    df = df.sort_values(["benchmark", "language", "backend"])
-
-    for benchmark, group in df.groupby("benchmark"):
-        print(f"### {str(benchmark).capitalize()} Benchmark")
-        param_cols = [
-            col.replace("parameters.", "")
-            for col in df.columns
-            if col.startswith("parameters.")
-        ]
-        display_cols = (
-            ["language", "backend"]
-            + [f"parameters.{p}" for p in param_cols]
-            + ["stats.mean", "stats.stddev", "stats.iterations"]
-        )
-        display_cols = [col for col in display_cols if col in group.columns]
-
-        group = group[display_cols].dropna(axis=1, how="all")
-
-        group.columns = [
-            col.replace("parameters.", "").replace("stats.", "")
-            for col in group.columns
-        ]
-
-        print(group.to_markdown(index=False))
-        print("\n")
+from compare_benches import compare_benchmarks
 
 
 def run_benchmarks():
